Perform type-conversion if we're running on a GPU
Basically, if we're running on a CPU we can use 64-bit precision,
however, if we're on a GPU we can only do 32-bit precision.  We had
hard-coded this and were starting to run into issues between the
two processing platforms.  This allows the code to perform a
one-time type-conversion (taking <1 sec) before training.